aqi_predictor/requirements.txt
# Essential Dependencies for AQI Prediction System

# Core Data Science Libraries
pandas>=1.5.0,<2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Machine Learning & Deep Learning
tensorflow>=2.13.0
xgboost>=1.7.0

# Feature Store & MLOps (with compatible versions)
hopsworks>=4.2.0,<4.3.0
urllib3>=1.25.4,<1.27

# Web Application
streamlit>=1.25.0

# Model Explainability
shap>=0.42.0

# API and Data Fetching
requests>=2.31.0
python-dotenv>=1.0.0

# Time Series & DateTime
pytz>=2023.3

# Data Processing
avro>=1.11.0

# Streaming (required by hopsworks)
confluent-kafka>=1.9.0

# Utilities
tqdm>=4.65.0
joblib>=1.3.0
pyyaml>=6.0

# Configuration
click>=8.1.0
